feat: metadata
