+++ app/user/[username]/page.tsx
@@ -1,11 +1,7 @@
-'use client'
 
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import Projects from "@/components/user/Portfolio/Projects";
-import Stats from "@/components/user/Portfolio/Stats";
-import HeaderBar from "@/components/user/userHeaderBar";
-import { useGetUser } from "@/hooks/user/getuser-username";
-import { Loader2 } from "lucide-react";
+import { User } from "@/types/User";
+import axios from "axios";
+import UserProfile from "./UserProfile";
 
 interface Props {
     params: {
@@ -13,36 +9,40 @@ interface Props {
     }
 }
 
-function ProfilePage({ params }: Props) {
-    const { data: user, isLoading } = useGetUser(params.username);
-    if (isLoading) {
-        return (
-            <div className="flex flex-col items-center justify-center w-full min-h-screen">
-                <Loader2 className="w-10 h-10 animate-spin" />
-            </div>
-        )
+
+export async function generateStaticParams() {
+    const { data } = await axios.get(`${process.env.NEXT_PUBLIC_BACKEND_URL}/getalluserusername`)
+    return data.users.map((user: any) => ({
+        params: {
+            username: user.username
+        }
+    }))
+}
+
+export async function generateMetadata({ params }: Props) {
+    const { data } = await axios.get(`${process.env.NEXT_PUBLIC_BACKEND_URL}/getalluserusername`)
+    const user: User = data.users.find((user: any) => user.username === params.username)
+    return {
+        title: user?.display_name || user?.username,
+        description: user?.bio,
+        type: "profile",
+        keywords: [user?.username, user?.display_name, user?.bio],
+        openGraph: {
+            title: user?.display_name || user?.username,
+            description: user?.bio,
+            type: "profile",
+            publishedTime: user?.created_at,
+            authors: [user?.username],
+        },
     }
-    if (!user) return (
-        <div className="flex flex-col items-center justify-center w-full min-h-screen">
-            <span>User Not Found</span>
-        </div>
-    )
+}
+
+
+function ProfilePage({ params }: Props) {
+
 
     return (
-        <div className="container min-h-screen mt-10">
-            <HeaderBar user={user?.data!} />
-            <Tabs defaultValue="portfolio" className="w-full min-h-screen mt-10">
-                <TabsList className="grid w-full grid-cols-2 gap-3">
-                    <TabsTrigger value="portfolio">Portfolio</TabsTrigger>
-                    <TabsTrigger value="resume">Resume</TabsTrigger>
-                </TabsList>
-                <TabsContent value="portfolio">
-                    <Stats karma={user.data.karma} />
-                    <Projects username={user.data.username} />
-                </TabsContent>
-                <TabsContent value="resume">Change your password here.</TabsContent>
-            </Tabs>
-        </div>
+        <UserProfile username={params.username} />
     );
 }
